Release v2.2.0 (Synapse)

+++ bazarr/utilities/backup.py
@@ -4,6 +4,7 @@ import os
 import sqlite3
 import shutil
 import logging
+import re
 
 from datetime import datetime, timedelta, timezone
 from zipfile import ZipFile, BadZipFile, ZIP_DEFLATED
@@ -15,12 +16,36 @@ from app.event_handler import event_stream
 from app.jobs_queue import jobs_queue
 from utilities.central import restart_bazarr
 
+_BACKUP_FILENAME_RE = re.compile(r'^bazarr_backup_v[\w.\-]+\.zip$')
+
+
+def _validate_backup_filename(filename):
+    """Return a safe filename if `filename` is a plain backup file name, else None."""
+    if not filename:
+        return None
+    safe_name = os.path.basename(filename)
+    if safe_name != filename:
+        return None
+    if not _BACKUP_FILENAME_RE.match(safe_name):
+        return None
+    return safe_name
+
+
+def _safe_extract(zip_obj, dest_path):
+    """Extract a ZIP archive while rejecting entries that escape `dest_path`."""
+    dest_real = os.path.realpath(dest_path)
+    for member in zip_obj.namelist():
+        member_real = os.path.realpath(os.path.join(dest_real, member))
+        if member_real != dest_real and not member_real.startswith(dest_real + os.sep):
+            raise BadZipFile(f'Unsafe path in backup archive: {member}')
+    zip_obj.extractall(path=dest_path)
+
 
 def get_backup_path():
     backup_dir = settings.backup.folder
     if not os.path.isdir(backup_dir):
         os.makedirs(backup_dir)
-    logging.debug(f'Backup directory path is: {backup_dir}')
+    logging.debug(f'Backup directory path is: {backup_dir}')  # noqa: G004
     return backup_dir
 
 
@@ -28,7 +53,7 @@ def get_restore_path():
     restore_dir = os.path.join(args.config_dir, 'restore')
     if not os.path.isdir(restore_dir):
         os.makedirs(restore_dir)
-    logging.debug(f'Restore directory path is: {restore_dir}')
+    logging.debug(f'Restore directory path is: {restore_dir}')  # noqa: G004
     return restore_dir
 
 
@@ -57,11 +82,11 @@ def backup_to_zip(job_id=None, wait_for_completion=False):
     database_backup_file = None
     now_string = now.strftime("%Y.%m.%d_%H.%M.%S")
     backup_filename = f"bazarr_backup_v{os.environ['BAZARR_VERSION']}_{now_string}.zip"
-    logging.debug(f'Backup filename will be: {backup_filename}')
+    logging.debug(f'Backup filename will be: {backup_filename}')  # noqa: G004
 
     if not settings.postgresql.enabled:
         database_src_file = os.path.join(args.config_dir, 'db', 'bazarr.db')
-        logging.debug(f'Database file path to backup is: {database_src_file}')
+        logging.debug(f'Database file path to backup is: {database_src_file}')  # noqa: G004
 
         try:
             database_src_con = sqlite3.connect(database_src_file)
@@ -79,7 +104,7 @@ def backup_to_zip(job_id=None, wait_for_completion=False):
             logging.exception('Unable to backup database file.')
 
     config_file = os.path.join(args.config_dir, 'config', 'config.yaml')
-    logging.debug(f'Config file path to backup is: {config_file}')
+    logging.debug(f'Config file path to backup is: {config_file}')  # noqa: G004
 
     with ZipFile(os.path.join(get_backup_path(), backup_filename), 'w', compression=ZIP_DEFLATED,
                  compresslevel=9) as backupZip:
@@ -88,7 +113,7 @@ def backup_to_zip(job_id=None, wait_for_completion=False):
             try:
                 os.remove(database_backup_file)
             except (OSError, FileNotFoundError):
-                logging.exception(f'Unable to delete temporary database backup file: {database_backup_file}')
+                logging.exception(f'Unable to delete temporary database backup file: {database_backup_file}')  # noqa: G004
         else:
             logging.debug('Database file is not included in backup. See previous exception')
         backupZip.write(config_file, 'config.yaml')
@@ -115,7 +140,7 @@ def restore_from_backup():
             shutil.copy(restore_config_path, dest_config_path)
             os.remove(restore_config_path)
         except (OSError, FileNotFoundError):
-            logging.exception(f'Unable to restore or delete config file to {dest_config_path}')
+            logging.exception(f'Unable to restore or delete config file to {dest_config_path}')  # noqa: G004
         else:
             if new_config:
                 if os.path.isfile(os.path.join(get_restore_path(), 'config.ini')):
@@ -127,7 +152,7 @@ def restore_from_backup():
             try:
                 shutil.copy(restore_database_path, dest_database_path)
             except (OSError, FileNotFoundError):
-                logging.exception(f'Unable to restore or delete db to {dest_database_path}')
+                logging.exception(f'Unable to restore or delete db to {dest_database_path}')  # noqa: G004
             else:
                 try:
                     if os.path.isfile(f'{dest_database_path}-shm'):
@@ -140,7 +165,7 @@ def restore_from_backup():
             try:
                 os.remove(restore_database_path)
             except (OSError, FileNotFoundError):
-                logging.exception(f'Unable to delete {dest_database_path}')
+                logging.exception(f'Unable to delete {dest_database_path}')  # noqa: G004
 
         logging.info('Backup restored successfully. Bazarr will restart.')
         from app.server import webserver
@@ -158,30 +183,34 @@ def restore_from_backup():
     except FileNotFoundError:
         pass
     except (OSError, FileNotFoundError):
-        logging.exception(f'Unable to delete {dest_config_path}')
+        logging.exception(f'Unable to delete {dest_config_path}')  # noqa: G004
 
 
 def prepare_restore(filename):
+    filename = _validate_backup_filename(filename)
+    if filename is None:
+        logging.error('Invalid backup filename refused for restore.')
+        return False
     src_zip_file_path = os.path.join(get_backup_path(), filename)
     dest_zip_file_path = os.path.join(get_restore_path(), filename)
     success = False
     try:
         shutil.copy(src_zip_file_path, dest_zip_file_path)
     except (OSError, FileNotFoundError):
-        logging.exception(f'Unable to copy backup archive to {dest_zip_file_path}')
+        logging.exception(f'Unable to copy backup archive to {dest_zip_file_path}')  # noqa: G004
     else:
         try:
             with ZipFile(dest_zip_file_path, 'r') as zipObj:
-                zipObj.extractall(path=get_restore_path())
+                _safe_extract(zipObj, get_restore_path())
         except BadZipFile:
-            logging.exception(f'Unable to extract files from backup archive {dest_zip_file_path}')
+            logging.exception(f'Unable to extract files from backup archive {dest_zip_file_path}')  # noqa: G004
         else:
             success = True
     finally:
         try:
             os.remove(dest_zip_file_path)
         except (OSError, FileNotFoundError):
-            logging.exception(f'Unable to delete backup archive {dest_zip_file_path}')
+            logging.exception(f'Unable to delete backup archive {dest_zip_file_path}')  # noqa: G004
 
     if success:
         logging.debug('time to restart')
@@ -203,25 +232,29 @@ def backup_rotation():
 
     backup_files = get_backup_files()
 
-    logging.debug(f'Cleaning up backup files older than {backup_retention} days')
+    logging.debug(f'Cleaning up backup files older than {backup_retention} days')  # noqa: G004
     for file in backup_files:
         if (datetime.fromtimestamp(os.path.getmtime(file), tz=timezone.utc) + timedelta(days=int(backup_retention)) <
                 datetime.now(tz=timezone.utc)):
-            logging.debug(f'Deleting old backup file {file}')
+            logging.debug(f'Deleting old backup file {file}')  # noqa: G004
             try:
                 os.remove(file)
             except (OSError, FileNotFoundError):
-                logging.debug(f'Unable to delete backup file {file}')
+                logging.debug(f'Unable to delete backup file {file}')  # noqa: G004
     logging.debug('Finished cleaning up old backup files')
 
 
 def delete_backup_file(filename):
+    filename = _validate_backup_filename(filename)
+    if filename is None:
+        logging.error('Invalid backup filename refused for deletion.')
+        return False
     backup_file_path = os.path.join(get_backup_path(), filename)
     try:
         os.remove(backup_file_path)
         return True
     except (OSError, FileNotFoundError):
-        logging.debug(f'Unable to delete backup file {backup_file_path}')
+        logging.debug(f'Unable to delete backup file {backup_file_path}')  # noqa: G004
     return False
